+++ src/main/java/org/javlo/context/GlobalContext.java
@@ -65,6 +65,7 @@ import org.javlo.helper.DebugHelper;
 import org.javlo.helper.ElementaryURLHelper;
 import org.javlo.helper.ElementaryURLHelper.Code;
 import org.javlo.helper.LangHelper;
+import org.javlo.helper.LocalLogger;
 import org.javlo.helper.NavigationHelper;
 import org.javlo.helper.ResourceHelper;
 import org.javlo.helper.ServletHelper;
@@ -1784,7 +1785,7 @@ public class GlobalContext implements Serializable, IPrintInfo {
 	public MenuElement getPageIfExist(ContentContext ctx, String url, boolean useURLCreator) throws Exception {
 		IURLFactory urlCreator = getURLFactory(ctx);
 		Map<String, MenuElement> localViewPages = viewPages;
-		if (ctx.getRenderMode() == ContentContext.VIEW_MODE && urlCreator != null && useURLCreator) {			
+		if (ctx.getRenderMode() == ContentContext.VIEW_MODE && urlCreator != null && useURLCreator) {
 			if (urlFromFactoryImported != urlCreator) {
 				synchronized (this.getLockLoadContent()) {
 					if (urlFromFactoryImported != urlCreator) {